Bugfix and clearer intRad computation
Bugfix: aggWithDelay... did not accept d=1 as viable parameter - fixed
Bugfix: Distance scripts had wrong argument syntax - fixed
just for aggWithDelaySpeedup:
- computation of intRad made more clear - was from intRad_1, now from intRad_2
- default stepPlotMod set to 5
- added "NoBoundary" as boundary conditions - makes experiment without boundaries, but starts in the [0,1]^d square as usual (also updated getInitHist script)

+++ MATLAB-code/auxiliaryFunctions/distances.m
@@ -13,8 +13,8 @@ function D = distances(x_1,x_2)
 %       element D(i,j) is distance ||x_1(i,:) - x_2(j,:)||
 
 arguments
-    x_1 (:,:) float
-    x_2 (:,:) float = x_1 
+    x_1 (:,:) double
+    x_2 (:,:) double = x_1 
 end
 
 if size(x_2) ~= size(x_1)
